requirements.txt
python-telegram-bot==21.6
asyncpg==0.29.0
python-dotenv==1.0.1
APScheduler==3.10.4
pytz==2024.2
greenlet==3.0.3
